refactor(main): separate cli run in app.go
