+++ .gitignore
@@ -28,7 +28,7 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
-fl
+./fl
 .task
 vendor/
 bin/
